+++ lib/core/networking/api_endpoints.dart
@@ -11,4 +11,10 @@ class ApiEndPoints {
 
   static String movieDetails(int movieId) => '/movie/$movieId';
   static String movieReviews(int movieId) => '/movie/$movieId/reviews';
+  static String movieCredits({
+    required int seriesId,
+    required int seasonNumber,
+    required int episodeNumber,
+  }) =>
+      '/tv/$seriesId/season/$seasonNumber/episode/$episodeNumber/credits';
 }
